requirements_txt.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0

# AI Provider SDKs
openai>=1.10.0
anthropic>=0.18.0
together>=0.2.7

# Optional dependencies for enhanced functionality
pandas>=2.0.0
numpy>=1.24.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0